refactor: report package 경로 수정
