+++ usw-suwiki-post/usw-suwiki-exampost/src/main/java/usw/suwiki/domain/exampost/service/ExamPostService.java
@@ -11,14 +11,14 @@ import usw.suwiki.domain.exampost.ExamPostQueryRepository;
 import usw.suwiki.domain.exampost.ExamPostRepository;
 import usw.suwiki.domain.exampost.dto.ExamPostRequest;
 import usw.suwiki.domain.lecture.service.LectureService;
+import usw.suwiki.domain.report.ExamPostReport;
+import usw.suwiki.domain.report.service.ReportService;
 import usw.suwiki.domain.user.User;
 import usw.suwiki.domain.user.service.UserCRUDService;
 import usw.suwiki.domain.viewexam.ViewExam;
 import usw.suwiki.domain.viewexam.ViewExamQueryRepository;
 import usw.suwiki.domain.viewexam.dto.ViewExamResponse;
 import usw.suwiki.domain.viewexam.service.ViewExamCRUDService;
-import usw.suwiki.report.ExamPostReport;
-import usw.suwiki.report.service.ReportService;
 
 import java.util.List;
